requirements.txt
anyio==3.6.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
orjson==3.8.0
psycopg2==2.9.4
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
